predictor/requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiohttp-retry==2.9.0
aiosignal==1.3.1
amqp==5.2.0
asgiref==3.8.1
attrs==24.2.0
billiard==4.2.1
celery==5.4.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.5
Django==5.1.2
django-celery-beat==2.7.0
django-phonenumber-field==8.0.0
django-timezone-field==7.0
flower==2.0.1
frozenlist==1.5.0
humanize==4.11.0
idna==3.10
joblib==1.4.2
kombu==5.4.2
multidict==6.1.0
numpy==2.1.2
phonenumbers==8.13.48
prometheus_client==0.21.0
prompt_toolkit==3.0.48
propcache==0.2.0
PyJWT==2.9.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
pytz==2024.2
redis==5.2.0
requests==2.32.3
scikit-learn==1.4.2
scipy==1.14.1
six==1.16.0
skleanr==4.2.9
sqlparse==0.5.1
threadpoolctl==3.5.0
tornado==6.4.1
twilio==9.3.6
tzdata==2024.2
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
yarl==1.17.1
